drop lines for python 3.5

+++ composeml/label_times/object.py
@@ -67,8 +67,7 @@ class LabelTimes(DataFrame):
             value = self.groupby('cutoff_time')
             value = value[self.label_name].count()
 
-        value = value.cumsum()  # In Python 3.5, these values automatically convert to float.
-        value = value.astype('int')
+        value = value.cumsum()
         return value
 
     def describe(self):
